+++ homeassistant/components/synology_dsm/camera.py
@@ -21,15 +21,15 @@ from homeassistant.helpers.entity import DeviceInfo
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 
-from . import SynoApi, SynologyDSMBaseEntity
+from . import SynoApi
 from .const import (
     CONF_SNAPSHOT_QUALITY,
     COORDINATOR_CAMERAS,
     DEFAULT_SNAPSHOT_QUALITY,
     DOMAIN,
     SYNO_API,
-    SynologyDSMEntityDescription,
 )
+from .models import SynologyDSMBaseEntity, SynologyDSMEntityDescription
 
 _LOGGER = logging.getLogger(__name__)
